:fire: remove unnecessary bin files
