--- bin/gh
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-const { openGitHub } = require("../lib");
-
-openGitHub();
