+++ bin/tbs-dock
@@ -3,9 +3,7 @@ var electron = require("electron");
 var spawn = require("cross-spawn");
 var argv = process.argv.slice(2);
 
-var result = spawn.sync(
-  electron,
-  [require.resolve("../lib/touchbar")].concat(argv),
-  { stdio: "ignore" }
-);
+var result = spawn.sync(electron, [require.resolve("../lib")].concat(argv), {
+  stdio: "ignore"
+});
 process.exit(result.status);
