--- bin/td
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-const { openTweetDeck } = require("../lib");
-
-openTweetDeck();
