--- bin/tw
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-const { openTwitter } = require("../lib");
-
-openTwitter();
