--- bin/yt
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-const { openYouTube } = require("../lib");
-
-openYouTube();
